refactor(web): extract the Explorer's tree controller into useFilesTree
FilesPanel mixed the fetch/refresh controller (per-directory correlation
ids, the open-panel prefetch, the refresh boundary, bell coalescing,
expand/collapse) with the panel frame, the file view, and the enlarge
state. The controller now lives in use-files-tree.ts beside the pure
store, mirroring how Changes splits use-changes-controller.ts from its
chrome; the panel keeps the frame and the file view. Behavior and DOM
are unchanged.
